Moved API calls to separate module and monkey patched gevent
