+++ controllers/bot/stakeout.py
@@ -20,6 +20,7 @@ from flask import flash, jsonify, redirect, render_template, request
 from flask_login import current_user, fresh_login_required, login_required
 
 import tasks
+import tasks.api
 import utils
 from models.faction import Faction
 from models.factionstakeoutmodel import FactionStakeoutModel
@@ -80,7 +81,7 @@ def stakeouts_dashboard(guildid: str):
                         f'[{stakeout.data["ID"]}] by the Tornium bot.',
                     }  # TODO: Add permission overwrite: everyone write false
 
-                channel = tasks.discordpost(f"guilds/{guildid}/channels", payload=payload)
+                channel = tasks.api.discordpost(f"guilds/{guildid}/channels", payload=payload)
 
                 db_stakeout = FactionStakeoutModel.objects(tid=request.form.get("factionid")).first()
                 db_stakeout.guilds[str(guildid)]["channel"] = int(channel["id"])
@@ -98,7 +99,7 @@ def stakeouts_dashboard(guildid: str):
                         }
                     ]
                 }
-                tasks.discordpost(
+                tasks.api.discordpost(
                     f'channels/{channel["id"]}/messages',
                     payload=message_payload,
                 )
@@ -134,7 +135,7 @@ def stakeouts_dashboard(guildid: str):
                         f'[{stakeout.data["player_id"]}] by the Tornium bot.',
                     }  # TODO: Add permission overwrite: everyone write false
 
-                channel = tasks.discordpost(f"guilds/{guildid}/channels", payload=payload)
+                channel = tasks.api.discordpost(f"guilds/{guildid}/channels", payload=payload)
 
                 db_stakeout = UserStakeoutModel.objects(tid=request.form.get("userid")).first()
                 db_stakeout.guilds[str(guildid)]["channel"] = int(channel["id"])
@@ -152,7 +153,7 @@ def stakeouts_dashboard(guildid: str):
                         }
                     ]
                 }
-                tasks.discordpost(
+                tasks.api.discordpost(
                     f'channels/{channel["id"]}/messages',
                     payload=message_payload,
                 )
@@ -336,7 +337,7 @@ def stakeout_update(guildid):
 
             stakeout = FactionStakeoutModel.objects(tid=faction).first()
             try:
-                tasks.discorddelete(f'channels/{stakeout.guilds[str(guildid)]["channel"]}')
+                tasks.api.discorddelete(f'channels/{stakeout.guilds[str(guildid)]["channel"]}')
             except utils.DiscordError as e:
                 if e.code == 10003:
                     pass
@@ -353,7 +354,7 @@ def stakeout_update(guildid):
 
             stakeout = UserStakeoutModel.objects(tid=user).first()
             try:
-                tasks.discorddelete(
+                tasks.api.discorddelete(
                     f'channels/{stakeout.guilds[str(guildid)]["channel"]}',
                 )
             except utils.DiscordError as e:
