+++ pyproject.toml
@@ -30,6 +30,7 @@ dependencies = [
     "flask ~= 2.2.2",
     "flask-cors ~= 3.0.10",
     "flask-login ~= 0.6.2",
+    "gevent ~= 22.10.2",
     "mongoengine ~= 0.26.0",
     "pynacl ~= 1.5.0",
     "redis ~= 4.5.1",
