+++ skynet/commands/faction/fulfill.py
@@ -17,6 +17,7 @@ import datetime
 import random
 
 import tasks
+import tasks.api
 import utils
 from models.faction import Faction
 from models.server import Server
@@ -257,7 +258,7 @@ def fulfill_command(interaction, *args, **kwargs):
         }
 
     try:
-        tasks.discordpatch(
+        tasks.api.discordpatch(
             f"channels/{faction.vault_config['banking']}/messages/{withdrawal.withdrawal_message}",
             {
                 "embeds": [
@@ -559,7 +560,7 @@ def fulfill_button(interaction, *args, **kwargs):
         }
 
     try:
-        tasks.discordpatch(
+        tasks.api.discordpatch(
             f"channels/{faction.vault_config['banking']}/messages/{withdrawal.withdrawal_message}",
             {
                 "embeds": [
